+++ CHANGELOG.md
@@ -17,6 +17,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ## [0.1.x]
 
+- Added first working low-level simple wrapper using pbllama 
+- Added first working high-level simple wrapper using pbllama
+
 - Added `nbllama.cpp` nanobind wrapper.
 - Added `pbllama.cpp` pybind11 wrapper.
 - Added `llama.pxd` file for `cyllama`